[CCUBE-1353] Removed unecessary styles, renamed modals state, moved the scrollableContainer to the renderModals function.
