+++ src/context-providers/display/modal-hook.tsx
@@ -27,7 +27,7 @@ export const useModal = () => {
     );
 
     return {
-        modals: state.modalState,
+        modals: state.modals,
         showModal,
         hideModal,
     };
